requirements.txt
crewai==0.11.2
openai>=1.7.1
pyyaml
